refactor: split documentation.ts into ui/webview.ts ## Summary Extract Webview HTML generation functions from documentation.ts to ui/webview.ts for better maintainability. ## Changes - Create src/ui/webview.ts (369 lines)   - generateNonce: CSP nonce generation   - buildCsp: Content Security Policy builder   - getDocumentHtml: Main document template   - getLoadingHtml: Loading state template   - getErrorHtml: Error state template - Update documentation.ts (378 lines, was 712)   - Import from ui/webview.ts   - Re-export for backward compatibility
🖥️ IDE: [Cursor](https://cursor.sh)
🔌 Extension: [Claude Code](https://claude.ai/download)

Model-Raw: claude-opus-4-5-20251101
